From: Paul Howarth <paul@city-fan.org>
Some of the GPL boilerplate refers to the Free Software Foundation's
address, where you can write to get a copy of the license. The FSF has
changed address (see http://www.fsf.org/about/contact/) so the
attached patch corrects references to the old address so that they now
point to the current one.

In doing so, I also took the liberty of updating the COPYING file to
be a copy of http://www.gnu.org/licenses/old-licenses/gpl-2.0.txt,
which is linked from
http://www.gnu.org/licenses/old-licenses/gpl-2.0.html. That fixes the
address and has some other minor wording and whitespace changes, such
as referring to the LGPL as "Lesser" rather than "Library".