Date: Tue, 6 Dec 2011 23:01:57 +0000 From: Paul Howarth <paul@city-fan.org> To: pptpclient-devel@lists.sourceforge.net Subject: [pptp-devel] [PATCH] fix comparisons of signed and unsigned integers
This patch fixes the warnings about comparisons of signed and unsigned
integers. I've assumed that packet sequence numbers really are unsigned
and that the code that deals with wrap-around works accordingly, which
I'm not completely sure about. I've also used the "z" printf format
modifier for printing ssize_t values, which is in C99 and POSIX
(http://pubs.opengroup.org/onlinepubs/009695399/functions/fprintf.html)
but might be missing for some systems?