pptp_callmgr: fix compiler warnings, #28
Reported in SourceForge bug #28 are two compiler warnings:
[pptp_callmgr.c:416]: (error) Dangerous usage of 'localaddr' (strncpy doesn't
always null-terminate it).
[pptp_callmgr.c:416]: (error) Dangerous usage of 'remoteaddr' (strncpy doesn't
always null-terminate it).

Since the source string is from a trusted source, inet_ntoa(3), it
cannot be more than 15 characters, so there is no impact.

In order to remove the warning, the code is changed to use strcpy(3)
instead of strncpy(3), and to use only one buffer instead of two.  A
buffer is needed because inet_ntoa(3) returns a pointer to a static
allocation that will be overwritten in the next call, and we make two
calls in the function.