Feature: terminate-from
To add a PPTP client to a Cisco VPDN group, terminate-from is often used
to specify the hostname of the L2TP access concentrator or L2TP network
server that will be required when accepting a VPDN tunnel.

That hostname is part of a PPTP control packet header and in the case of
Linux PPTP client, hardcoded to "local". This patch provides the ability
to specify a different one through a command line argument.